+++ bin/gtfs-to-geojson.js
@@ -2,10 +2,13 @@
 
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
+import PrettyError from 'pretty-error';
 
 import gtfsToGeoJSON from '../lib/gtfs-to-geojson.js';
 import { getConfig } from '../lib/file-utils.js';
 import { formatError } from '../lib/log-utils.js';
+
+const pe = new PrettyError();
 
 const { argv } = yargs(hideBin(process.argv))
   .usage('Usage: $0 --config ./config.json')
@@ -14,19 +17,19 @@ const { argv } = yargs(hideBin(process.argv))
     alias: 'configPath',
     describe: 'Path to config file',
     default: './config.json',
-    type: 'string'
+    type: 'string',
   })
   .option('s', {
     alias: 'skipImport',
-    describe: 'Don\'t import GTFS file.',
-    type: 'boolean'
+    describe: "Don't import GTFS file.",
+    type: 'boolean',
   })
   .default('skipImport', undefined);
 
-const handleError = error => {
+const handleError = (error) => {
   const text = error || 'Unknown Error';
   process.stdout.write(`\n${formatError(text)}\n`);
-  console.error(error);
+  console.error(pe.render(error));
   process.exit(1);
 };
 
@@ -36,5 +39,4 @@ const setupImport = async () => {
   process.exit();
 };
 
-setupImport()
-  .catch(handleError);
+setupImport().catch(handleError);